+++ src/main/java/com/sparta/hhplusconcert/concert/api/response/GetConcertDatesResponse.java
@@ -1,4 +1,4 @@
-package com.sparta.hhplusconcert.api.concert.response;
+package com.sparta.hhplusconcert.concert.api.response;
 
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
