+++ src/main/java/com/sparta/hhplusconcert/concert/domain/ReservationStatus.java
@@ -1,4 +1,4 @@
-package com.sparta.hhplusconcert.domain.concert;
+package com.sparta.hhplusconcert.concert.domain;
 
 public enum ReservationStatus {
   PENDING_PAYMENT,
