+++ src/main/java/com/sparta/hhplusconcert/concert/domain/entity/ConcertSeatEntity.java
@@ -1,26 +1,30 @@
-package com.sparta.hhplusconcert.domain.concert.entity;
+package com.sparta.hhplusconcert.concert.domain.entity;
 
 import com.sparta.hhplusconcert.domain.common.TimeBaseEntity;
+import com.sparta.hhplusconcert.concert.domain.SeatStatus;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import java.time.LocalDate;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Entity
-@Table(name = "CONCERT_SCHEDULE")
+@Table(name = "CONCERT_SEAT")
 @Getter
+@Setter
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class ConcertScheduleEntity extends TimeBaseEntity {
+public class ConcertSeatEntity extends TimeBaseEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
@@ -29,11 +33,12 @@ public class ConcertScheduleEntity extends TimeBaseEntity {
   private Long concertId;
 
   @Column(nullable = false)
-  private LocalDate date;
-
-  private Integer seatCapacity = 50;
+  private Long scheduleId;
 
   @Column(nullable = false)
-  private Integer seatLeft;
+  private Integer seatNumber;
+
+  @Enumerated(EnumType.STRING)
+  private SeatStatus status;
 
 }
