+++ src/main/java/com/sparta/hhplusconcert/concert/infra/ConcertScheduleRepositoryImpl.java
@@ -1,6 +1,6 @@
-package com.sparta.hhplusconcert.infra.concert;
+package com.sparta.hhplusconcert.concert.infra;
 
-import com.sparta.hhplusconcert.domain.concert.entity.ConcertScheduleEntity;
+import com.sparta.hhplusconcert.concert.domain.entity.ConcertScheduleEntity;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
