+++ src/main/java/com/sparta/hhplusconcert/concert/infra/ConcertSeatRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.hhplusconcert.infra.concert;
+package com.sparta.hhplusconcert.concert.infra;
 
-import com.sparta.hhplusconcert.domain.concert.entity.ConcertSeatEntity;
+import com.sparta.hhplusconcert.concert.domain.entity.ConcertSeatEntity;
 import java.util.List;
 
 public interface ConcertSeatRepository {
